Modfiy paging state setter and move to statement.rs
Paging state bytes are set without being cast to u8.
Previously the setter added a null-terminator which is not its original
and intended behavior.
The setter is moved to statement.rs as it is a statement-related
function.
